requirements.txt
# UI
streamlit
pandas

# Database
neo4j
python-dotenv

# Agent + MCP
openai
openai-agents
mcp
